- float values changed to double in NNet, Dependency, Neuron - Middle Neurons should be working correctly now - Added AdamZero - Added a NNet file

+++ AI/Dependency.cs
@@ -9,6 +9,6 @@ namespace ChaosTerraria.AI
     public class Dependency
     {
         public string neuronId;
-        public float weight;
+        public double weight;
     }
 }
